requirements.txt
easysnmp==0.2.5
PyYAML==5.4.1
